feat(frontend): add token picker and wire i18n modal accessibility
Closes #607
Closes #609
Closes #611
Closes #613
